@@ -1,10 +1,8 @@
 ﻿namespace Ardalis.SmartEnum.Benchmarks
 {
-    using System;
-    using BenchmarkDotNet.Attributes;
     using BenchmarkDotNet.Running;
 
-    public class Program
+    public static class Program
     {
         public static void Main(string[] args)
         {
