@@ -10,65 +10,65 @@ namespace Ardalis.SmartEnum.Benchmarks
         // Enum
 
         [Benchmark]
-        public TestEnum Enum_Switch() 
+        public TestEnum Enum_Switch()
         {
             var value = TestEnum.Ten;
-            switch(value)
+            switch (value)
             {
-                case TestEnum.One: 
+                case TestEnum.One:
                     return TestEnum.One;
-                case TestEnum.Two: 
+                case TestEnum.Two:
                     return TestEnum.Two;
-                case TestEnum.Three: 
+                case TestEnum.Three:
                     return TestEnum.Three;
-                case TestEnum.Four: 
+                case TestEnum.Four:
                     return TestEnum.Four;
-                case TestEnum.Five: 
+                case TestEnum.Five:
                     return TestEnum.Five;
-                case TestEnum.Six: 
+                case TestEnum.Six:
                     return TestEnum.Six;
-                case TestEnum.Seven: 
+                case TestEnum.Seven:
                     return TestEnum.Seven;
-                case TestEnum.Eight: 
+                case TestEnum.Eight:
                     return TestEnum.Eight;
-                case TestEnum.Nine: 
+                case TestEnum.Nine:
                     return TestEnum.Nine;
-                case TestEnum.Ten: 
+                case TestEnum.Ten:
                     return TestEnum.Ten;
                 default: throw new Exception();
             }
-        }  
+        }
 
         [Benchmark]
-        public TestEnum Enum_IfEqualsMethodCascade() 
+        public TestEnum Enum_IfEqualsMethodCascade()
         {
             var value = TestEnum.Ten;
-            if (value.Equals(TestEnum.One)) 
+            if (value.Equals(TestEnum.One))
                 return TestEnum.One;
-            if (value.Equals(TestEnum.Two)) 
+            if (value.Equals(TestEnum.Two))
                 return TestEnum.Two;
-            if (value.Equals(TestEnum.Three)) 
+            if (value.Equals(TestEnum.Three))
                 return TestEnum.Three;
-            if (value.Equals(TestEnum.Five)) 
+            if (value.Equals(TestEnum.Five))
                 return TestEnum.Five;
-            if (value.Equals(TestEnum.Six)) 
+            if (value.Equals(TestEnum.Six))
                 return TestEnum.Six;
-            if (value.Equals(TestEnum.Seven)) 
+            if (value.Equals(TestEnum.Seven))
                 return TestEnum.Seven;
-            if (value.Equals(TestEnum.Eight)) 
+            if (value.Equals(TestEnum.Eight))
                 return TestEnum.Eight;
-            if (value.Equals(TestEnum.Nine)) 
+            if (value.Equals(TestEnum.Nine))
                 return TestEnum.Nine;
-            if (value.Equals(TestEnum.Ten)) 
+            if (value.Equals(TestEnum.Ten))
                 return TestEnum.Ten;
             throw new Exception();
-        } 
+        }
 
         [Benchmark]
-        public TestEnum Enum_IfEqualOperatorCascade() 
+        public TestEnum Enum_IfEqualOperatorCascade()
         {
             var value = TestEnum.Ten;
-            if (value == TestEnum.One) 
+            if (value == TestEnum.One)
                 return TestEnum.One;
             if (value == TestEnum.Two)
                 return TestEnum.Two;
@@ -87,94 +87,94 @@ namespace Ardalis.SmartEnum.Benchmarks
             if (value == TestEnum.Ten)
                 return TestEnum.Ten;
             throw new Exception();
-        }  
+        }
 
         ////////////////////////////////////////////////////////////////////////////////
         // SmartEnum
 
         [Benchmark]
-        public TestSmartEnum SmartEnum_Switch() 
+        public TestSmartEnum SmartEnum_Switch()
         {
             var value = TestSmartEnum.Ten;
-            switch(value.Name)
+            switch (value.Name)
             {
-                case nameof(TestSmartEnum.One): 
+                case nameof(TestSmartEnum.One):
                     return TestSmartEnum.One;
-                case nameof(TestSmartEnum.Two): 
+                case nameof(TestSmartEnum.Two):
                     return TestSmartEnum.Two;
-                case nameof(TestSmartEnum.Three): 
+                case nameof(TestSmartEnum.Three):
                     return TestSmartEnum.Three;
-                case nameof(TestSmartEnum.Four): 
+                case nameof(TestSmartEnum.Four):
                     return TestSmartEnum.Four;
-                case nameof(TestSmartEnum.Five): 
+                case nameof(TestSmartEnum.Five):
                     return TestSmartEnum.Five;
-                case nameof(TestSmartEnum.Six): 
+                case nameof(TestSmartEnum.Six):
                     return TestSmartEnum.Six;
-                case nameof(TestSmartEnum.Seven): 
+                case nameof(TestSmartEnum.Seven):
                     return TestSmartEnum.Seven;
-                case nameof(TestSmartEnum.Eight): 
+                case nameof(TestSmartEnum.Eight):
                     return TestSmartEnum.Eight;
-                case nameof(TestSmartEnum.Nine): 
+                case nameof(TestSmartEnum.Nine):
                     return TestSmartEnum.Nine;
-                case nameof(TestSmartEnum.Ten): 
+                case nameof(TestSmartEnum.Ten):
                     return TestSmartEnum.Ten;
-                default: 
+                default:
                     throw new Exception();
             }
         }
 
         [Benchmark]
-        public TestSmartEnum SmartEnum_SwitchPatternMatching() 
+        public TestSmartEnum SmartEnum_SwitchPatternMatching()
         {
             var value = TestSmartEnum.Ten;
-            switch(value)
+            switch (value)
             {
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.One): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.One):
                     return TestSmartEnum.One;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Two): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Two):
                     return TestSmartEnum.Two;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Three):  
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Three):
                     return TestSmartEnum.Three;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Four): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Four):
                     return TestSmartEnum.Four;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Five): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Five):
                     return TestSmartEnum.Five;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Six): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Six):
                     return TestSmartEnum.Six;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Seven): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Seven):
                     return TestSmartEnum.Seven;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Eight): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Eight):
                     return TestSmartEnum.Eight;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Nine): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Nine):
                     return TestSmartEnum.Nine;
-                case var smartEnum when smartEnum.Equals(TestSmartEnum.Ten): 
+                case var smartEnum when smartEnum.Equals(TestSmartEnum.Ten):
                     return TestSmartEnum.Ten;
-                default: 
+                default:
                     throw new Exception();
             }
         }
 
         [Benchmark]
-        public TestSmartEnum SmartEnum_IfCascade() 
+        public TestSmartEnum SmartEnum_IfCascade()
         {
             var value = TestSmartEnum.Ten;
-            if (value.Equals(TestSmartEnum.One)) 
+            if (value.Equals(TestSmartEnum.One))
                 return TestSmartEnum.One;
-            if (value.Equals(TestSmartEnum.Two)) 
+            if (value.Equals(TestSmartEnum.Two))
                 return TestSmartEnum.Two;
-            if (value.Equals(TestSmartEnum.Three)) 
+            if (value.Equals(TestSmartEnum.Three))
                 return TestSmartEnum.Three;
-            if (value.Equals(TestSmartEnum.Five)) 
+            if (value.Equals(TestSmartEnum.Five))
                 return TestSmartEnum.Five;
-            if (value.Equals(TestSmartEnum.Six)) 
+            if (value.Equals(TestSmartEnum.Six))
                 return TestSmartEnum.Six;
-            if (value.Equals(TestSmartEnum.Seven)) 
+            if (value.Equals(TestSmartEnum.Seven))
                 return TestSmartEnum.Seven;
-            if (value.Equals(TestSmartEnum.Eight)) 
+            if (value.Equals(TestSmartEnum.Eight))
                 return TestSmartEnum.Eight;
-            if (value.Equals(TestSmartEnum.Nine)) 
+            if (value.Equals(TestSmartEnum.Nine))
                 return TestSmartEnum.Nine;
-            if (value.Equals(TestSmartEnum.Ten)) 
+            if (value.Equals(TestSmartEnum.Ten))
                 return TestSmartEnum.Ten;
             throw new Exception();
         }
