@@ -5,26 +5,11 @@ namespace Ardalis.SmartEnum.JsonNet
 
     public class SmartEnumValueConverter<TEnum, TValue> : JsonConverter<TEnum>
         where TEnum : SmartEnum<TEnum, TValue>
-        where TValue : struct, IEquatable<TValue>, IComparable<TValue>
+        where TValue : IEquatable<TValue>, IComparable<TValue>
     {
-        /// <summary>
-        /// 
-        /// </summary>
         public override bool CanRead => true;
-        /// <summary>
-        /// 
-        /// </summary>
         public override bool CanWrite => true;
 
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="reader"></param>
-        /// <param name="objectType"></param>
-        /// <param name="existingValue"></param>
-        /// <param name="hasExistingValue"></param>
-        /// <param name="serializer"></param>
-        /// <returns></returns>
         public override TEnum ReadJson(JsonReader reader, Type objectType, TEnum existingValue, bool hasExistingValue, JsonSerializer serializer)
         {
             try
@@ -47,12 +32,6 @@ namespace Ardalis.SmartEnum.JsonNet
             }
         }
 
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="writer"></param>
-        /// <param name="value"></param>
-        /// <param name="serializer"></param>
         public override void WriteJson(JsonWriter writer, TEnum value, JsonSerializer serializer)
         {
             if (value is null)
