@@ -4,25 +4,14 @@ namespace Ardalis.SmartEnum.MessagePack
     using global::MessagePack;
     using global::MessagePack.Formatters;
 
-    /// <summary>
-    /// 
-    /// </summary>
     public class SmartEnumNameResolver : IFormatterResolver
     {
-        /// <summary>
-        /// 
-        /// </summary>
         public static readonly SmartEnumNameResolver Instance = new SmartEnumNameResolver();
 
-        private SmartEnumNameResolver() 
+        private SmartEnumNameResolver()
         {
         }
 
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <returns></returns>
         public IMessagePackFormatter<T> GetFormatter<T>() =>
             FormatterCache<T>.Formatter;
 
@@ -36,7 +25,7 @@ namespace Ardalis.SmartEnum.MessagePack
                 if (typeof(T).IsSmartEnum(out var genericArguments))
                 {
                     var formatterType = typeof(SmartEnumNameFormatter<,>).MakeGenericType(genericArguments);
-                    Formatter = (IMessagePackFormatter<T>)Activator.CreateInstance(formatterType);                    
+                    Formatter = (IMessagePackFormatter<T>)Activator.CreateInstance(formatterType);
                 }
             }
 #pragma warning restore S3963 // "static" fields should be initialized inline
