@@ -1,32 +1,26 @@
 namespace Ardalis.SmartEnum.Utf8Json
 {
     using global::Utf8Json;
-    using global::Utf8Json.Internal;
     using System;
 
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <typeparam name="TEnum"></typeparam>
-    /// <typeparam name="TValue"></typeparam>
     public class SmartEnumNameFormatter<TEnum, TValue> : IJsonFormatter<TEnum>
         where TEnum : SmartEnum<TEnum, TValue>
         where TValue : struct, IEquatable<TValue>, IComparable<TValue>
     {
         public void Serialize(ref JsonWriter writer, TEnum value, IJsonFormatterResolver formatterResolver)
         {
-            if (value is null) 
-            { 
-                writer.WriteNull(); 
-                return; 
+            if (value is null)
+            {
+                writer.WriteNull();
+                return;
             }
 
-            writer.WriteString(value.Name); 
+            writer.WriteString(value.Name);
         }
 
         public TEnum Deserialize(ref JsonReader reader, IJsonFormatterResolver formatterResolver)
         {
-            if (reader.ReadIsNull()) 
+            if (reader.ReadIsNull())
                 return null;
 
             var name = reader.ReadString();
