@@ -1,7 +1,6 @@
 namespace Ardalis.SmartEnum.Utf8Json
 {
     using global::Utf8Json;
-    using global::Utf8Json.Internal;
     using System;
 
     public class SmartEnumValueFormatter<TEnum, TValue> : IJsonFormatter<TEnum>
@@ -10,33 +9,33 @@ namespace Ardalis.SmartEnum.Utf8Json
     {
         public void Serialize(ref JsonWriter writer, TEnum value, IJsonFormatterResolver formatterResolver)
         {
-            if (value is null) 
-            { 
-                writer.WriteNull(); 
-                return; 
+            if (value is null)
+            {
+                writer.WriteNull();
+                return;
             }
 
-            if(typeof(TValue) == typeof(bool))
+            if (typeof(TValue) == typeof(bool))
                 writer.WriteBoolean((bool)(object)value.Value);
-            else if(typeof(TValue) == typeof(byte))
+            else if (typeof(TValue) == typeof(byte))
                 writer.WriteByte((byte)(object)value.Value);
-            else if(typeof(TValue) == typeof(sbyte))
+            else if (typeof(TValue) == typeof(sbyte))
                 writer.WriteSByte((sbyte)(object)value.Value);
-            else if(typeof(TValue) == typeof(short))
+            else if (typeof(TValue) == typeof(short))
                 writer.WriteInt16((short)(object)value.Value);
-            else if(typeof(TValue) == typeof(ushort))
+            else if (typeof(TValue) == typeof(ushort))
                 writer.WriteUInt16((ushort)(object)value.Value);
-            else if(typeof(TValue) == typeof(int))
+            else if (typeof(TValue) == typeof(int))
                 writer.WriteInt32((int)(object)value.Value);
-            else if(typeof(TValue) == typeof(uint))
+            else if (typeof(TValue) == typeof(uint))
                 writer.WriteUInt32((uint)(object)value.Value);
-            else if(typeof(TValue) == typeof(long))
+            else if (typeof(TValue) == typeof(long))
                 writer.WriteInt64((long)(object)value.Value);
-            else if(typeof(TValue) == typeof(ulong))
+            else if (typeof(TValue) == typeof(ulong))
                 writer.WriteUInt64((ulong)(object)value.Value);
-            else if(typeof(TValue) == typeof(float))
+            else if (typeof(TValue) == typeof(float))
                 writer.WriteSingle((float)(object)value.Value);
-            else if(typeof(TValue) == typeof(double))
+            else if (typeof(TValue) == typeof(double))
                 writer.WriteDouble((double)(object)value.Value);
             else
                 throw new ArgumentOutOfRangeException(typeof(TValue).ToString(), $"{typeof(TValue).Name} is not supported.");
@@ -44,7 +43,7 @@ namespace Ardalis.SmartEnum.Utf8Json
 
         public TEnum Deserialize(ref JsonReader reader, IJsonFormatterResolver formatterResolver)
         {
-            if (reader.ReadIsNull()) 
+            if (reader.ReadIsNull())
                 return null;
 
             return SmartEnum<TEnum, TValue>.FromValue(ReadValue(ref reader));
@@ -52,27 +51,27 @@ namespace Ardalis.SmartEnum.Utf8Json
 
         TValue ReadValue(ref JsonReader reader)
         {
-            if(typeof(TValue) == typeof(bool))
+            if (typeof(TValue) == typeof(bool))
                 return (TValue)(object)reader.ReadBoolean();
-            if(typeof(TValue) == typeof(byte))
+            if (typeof(TValue) == typeof(byte))
                 return (TValue)(object)reader.ReadByte();
-            if(typeof(TValue) == typeof(sbyte))
+            if (typeof(TValue) == typeof(sbyte))
                 return (TValue)(object)reader.ReadSByte();
-            if(typeof(TValue) == typeof(short))
+            if (typeof(TValue) == typeof(short))
                 return (TValue)(object)reader.ReadInt16();
-            if(typeof(TValue) == typeof(ushort))
+            if (typeof(TValue) == typeof(ushort))
                 return (TValue)(object)reader.ReadUInt16();
-            if(typeof(TValue) == typeof(int))
+            if (typeof(TValue) == typeof(int))
                 return (TValue)(object)reader.ReadInt32();
-            if(typeof(TValue) == typeof(uint))
+            if (typeof(TValue) == typeof(uint))
                 return (TValue)(object)reader.ReadUInt32();
-            if(typeof(TValue) == typeof(long))
+            if (typeof(TValue) == typeof(long))
                 return (TValue)(object)reader.ReadInt64();
-            if(typeof(TValue) == typeof(ulong))
+            if (typeof(TValue) == typeof(ulong))
                 return (TValue)(object)reader.ReadUInt64();
-            if(typeof(TValue) == typeof(float))
+            if (typeof(TValue) == typeof(float))
                 return (TValue)(object)reader.ReadSingle();
-            if(typeof(TValue) == typeof(double))
+            if (typeof(TValue) == typeof(double))
                 return (TValue)(object)reader.ReadDouble();
             throw new ArgumentOutOfRangeException(typeof(TValue).ToString(), $"{typeof(TValue).Name} is not supported.");
         }
