@@ -6,6 +6,6 @@
         public static readonly TestEnum Two = new TestEnum(nameof(Two), 2);
         public static readonly TestEnum Three = new TestEnum(nameof(Three), 3);
 
-        TestEnum(string name, int value) : base(name, value) {}
+        TestEnum(string name, int value) : base(name, value) { }
     }
 }
