@@ -1,6 +1,5 @@
 namespace Ardalis.SmartEnum.ProtoBufNet
 {
-    using System;
     using System.IO;
     using ProtoBuf.Meta;
 
@@ -8,12 +7,12 @@ namespace Ardalis.SmartEnum.ProtoBufNet
     {
         public static T DeepClone<T>(T value, TypeModel model)
         {
-            using(var stream = new MemoryStream())
+            using (var stream = new MemoryStream())
             {
                 model.Serialize(stream, value);
                 stream.Seek(0, SeekOrigin.Begin);
                 return (T)model.Deserialize(stream, null, typeof(T));
-            }        
-        } 
+            }
+        }
     }
 }
