@@ -1,13 +1,12 @@
 namespace Ardalis.SmartEnum.UnitTests
 {
-    using System;
     using FluentAssertions;
     using Xunit;
 
     public class SmartEnumCompareTo
     {
         public static TheoryData<TestEnum, TestEnum, int> CompareToData =>
-            new TheoryData<TestEnum, TestEnum, int> 
+            new TheoryData<TestEnum, TestEnum, int>
             {
                 { TestEnum.Two, TestEnum.One, 1 },
                 { TestEnum.Two, TestEnum.Two, 0 },
@@ -19,12 +18,12 @@ namespace Ardalis.SmartEnum.UnitTests
         public void CompareToReturnsExpected(TestEnum left, TestEnum right, int expected)
         {
             var result = left.CompareTo(right);
-            
+
             result.Should().Be(expected);
         }
 
         public static TheoryData<TestEnum, TestEnum, bool, bool, bool> ComparisonOperatorsData =>
-            new TheoryData<TestEnum, TestEnum, bool, bool, bool> 
+            new TheoryData<TestEnum, TestEnum, bool, bool, bool>
             {
                 { TestEnum.Two, TestEnum.One, false, false, true },
                 { TestEnum.Two, TestEnum.Two, false, true, false },
@@ -67,8 +66,8 @@ namespace Ardalis.SmartEnum.UnitTests
             var result = left >= right;
 
             result.Should().Be(greaterThan || equalTo);
-        }    
+        }
 
         #pragma warning restore xUnit1026 // Theory method does not use parameter
     }
-} 
+}
