@@ -6,20 +6,20 @@ namespace Ardalis.SmartEnum.UnitTests
 
     public class SmartEnumExtensionsTests
     {
-        public abstract class AbstractEnum : SmartEnum<AbstractEnum, int> 
+        public abstract class AbstractEnum : SmartEnum<AbstractEnum, int>
         {
-            protected AbstractEnum(string name, int value) : base(name, value) {}
+            protected AbstractEnum(string name, int value) : base(name, value) { }
         }
 
-        public class GenericEnum<T> : 
-            SmartEnum<GenericEnum<T>, T> 
+        public class GenericEnum<T> :
+            SmartEnum<GenericEnum<T>, T>
             where T : struct, IEquatable<T>, IComparable<T>
         {
-            protected GenericEnum(string name, T value) : base(name, value) {}
+            protected GenericEnum(string name, T value) : base(name, value) { }
         }
 
         public static TheoryData<Type, bool, Type[]> IsSmartEnumData =>
-            new TheoryData<Type, bool, Type[]> 
+            new TheoryData<Type, bool, Type[]>
             {
                 { typeof(int), false, null },
                 { typeof(AbstractEnum), false, null },
@@ -45,7 +45,7 @@ namespace Ardalis.SmartEnum.UnitTests
             var result = type.IsSmartEnum(out var genericArguments);
 
             result.Should().Be(expectedResult);
-            if(result)
+            if (result)
             {
                 genericArguments.Should().Equal(expectedGenericArguments);
             }
