@@ -16,20 +16,21 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
 
             [JsonFormatter(typeof(SmartEnumNameFormatter<TestEnumInt16, short>))]
             public TestEnumInt16 Int16 { get; set; }
-            
+
             [JsonFormatter(typeof(SmartEnumNameFormatter<TestEnumInt32, int>))]
-            public TestEnumInt32 Int32 { get; set; }        
+            public TestEnumInt32 Int32 { get; set; }
 
             [JsonFormatter(typeof(SmartEnumNameFormatter<TestEnumDouble, double>))]
-            public TestEnumDouble Double { get; set; }        
+            public TestEnumDouble Double { get; set; }
         }
 
-        static readonly TestClass TestInstance = new TestClass { 
+        static readonly TestClass TestInstance = new TestClass
+        {
             Bool = TestEnumBoolean.Instance,
             Int16 = TestEnumInt16.Instance,
             Int32 = TestEnumInt32.Instance,
             Double = TestEnumDouble.Instance,
-         };
+        };
 
         static readonly string JsonString = @"{""Bool"":""Instance"",""Int16"":""Instance"",""Int32"":""Instance"",""Double"":""Instance""}";
 
@@ -40,7 +41,7 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
                 new SmartEnumNameFormatter<TestEnumInt16, short>(),
                 new SmartEnumNameFormatter<TestEnumInt32, int>(),
                 new SmartEnumNameFormatter<TestEnumDouble, double>()
-            );       
+            );
         }
 
         [Fact]
@@ -49,7 +50,7 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
             var json = JsonSerializer.Serialize(TestInstance);
 
             Encoding.UTF8.GetString(json).Should().Be(JsonString);
-         }
+        }
 
         [Fact]
         public void DeserializesNames()
@@ -60,7 +61,7 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
             obj.Int16.Should().BeSameAs(TestEnumInt16.Instance);
             obj.Int32.Should().BeSameAs(TestEnumInt32.Instance);
             obj.Double.Should().BeSameAs(TestEnumDouble.Instance);
-        }    
+        }
 
         [Fact]
         public void DeserializesNullByDefault()
@@ -73,7 +74,7 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
             obj.Int16.Should().BeNull();
             obj.Int32.Should().BeNull();
             obj.Double.Should().BeNull();
-        }    
+        }
 
 
         [Fact]
@@ -87,7 +88,7 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
             obj.Int16.Should().BeNull();
             obj.Int32.Should().BeNull();
             obj.Double.Should().BeNull();
-        }  
+        }
 
         [Fact]
         public void DeserializeThrowsWhenNotFound()
@@ -99,6 +100,6 @@ namespace Ardalis.SmartEnum.Utf8Json.UnitTests
             act.Should()
                 .Throw<SmartEnumNotFoundException>()
                 .WithMessage($@"No {nameof(TestEnumBoolean)} with Name ""Not Found"" found.");
-        }  
+        }
     }
 }
